feat: refactor dataset loading by introducing load_dataset_messages utility function
